TIP-1173: move class about atributes in cache namespace instead of SQL
