+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/queries.yml
@@ -5,11 +5,11 @@ services:
             $connection: '@database_connection'
 
     akeneo.pim.structure.query.sql_get_attributes:
-        class: 'Akeneo\Pim\Structure\Bundle\Query\PublicApi\AttributeTypes\Sql\SqlGetAttributes'
+        class: 'Akeneo\Pim\Structure\Bundle\Query\PublicApi\Attribute\Sql\SqlGetAttributes'
         arguments:
             $connection: '@database_connection'
 
     akeneo.pim.structure.query.get_attributes:
-        class: 'Akeneo\Pim\Structure\Bundle\Query\PublicApi\AttributeTypes\Sql\LRUCachedGetAttributes'
+        class: 'Akeneo\Pim\Structure\Bundle\Query\PublicApi\Attribute\Cache\LRUCachedGetAttributes'
         arguments:
             $getAttributes: '@akeneo.pim.structure.query.sql_get_attributes'
