+++ tests/back/Pim/Structure/Specification/Bundle/Query/PublicApi/Attribute/Cache/LRUCachedGetAttributesSpec.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Specification\Akeneo\Pim\Structure\Bundle\Query\PublicApi\AttributeTypes\Sql;
+namespace Specification\Akeneo\Pim\Structure\Bundle\Query\PublicApi\Attribute\Cache;
 
 use Akeneo\Pim\Structure\Component\AttributeTypes;
 use Akeneo\Pim\Structure\Component\Query\PublicApi\AttributeType\Attribute;
